+++ ase/test/constraints/test_fixsubsetcom.py
@@ -5,7 +5,7 @@ import pytest
 from ase import Atoms
 from ase.build import molecule
 from ase.calculators.emt import EMT
-from ase.constraints import FixCom
+from ase.constraints import FixSubsetCom
 from ase.md.velocitydistribution import MaxwellBoltzmannDistribution
 from ase.optimize import BFGS
 
@@ -13,8 +13,11 @@ from ase.optimize import BFGS
 @pytest.fixture(name="atoms")
 def fixture_atoms() -> Atoms:
     """fixture_atoms"""
-    atoms = molecule('H2O')
-    atoms.center(vacuum=4)
+    atoms0 = molecule('H2O')
+    atoms0.positions -= 2.0
+    atoms1 = molecule('H2O')
+    atoms1.positions += 2.0
+    atoms = atoms0 + atoms1
     atoms.calc = EMT()
     return atoms
 
@@ -22,13 +25,14 @@ def fixture_atoms() -> Atoms:
 @pytest.mark.optimize
 def test_center_of_mass_position(atoms: Atoms):
     """Test if the center of mass does not move."""
-    cold = atoms.get_center_of_mass()
-    atoms.set_constraint(FixCom())
+    indices = [3, 4, 5]
+    cold = atoms.get_center_of_mass(indices=indices)
+    atoms.set_constraint(FixSubsetCom(indices=indices))
 
     with BFGS(atoms) as opt:
         opt.run(steps=5)
 
-    cnew = atoms.get_center_of_mass()
+    cnew = atoms.get_center_of_mass(indices=indices)
 
     assert max(cnew - cold) == pytest.approx(0.0, abs=1e-8)
 
@@ -36,12 +40,15 @@ def test_center_of_mass_position(atoms: Atoms):
 @pytest.mark.optimize
 def test_center_of_mass_velocity(atoms: Atoms):
     """Test if the center-of-mass veloeicty is zero."""
-    atoms.set_constraint(FixCom())
+    indices = [3, 4, 5]
+    atoms.set_constraint(FixSubsetCom(indices=indices))
 
     # `adjust_momenta` of constaints are applied inside
     MaxwellBoltzmannDistribution(atoms, temperature_K=300.0)
 
-    velocity_com = atoms.get_momenta().sum(axis=0) / atoms.get_masses().sum()
+    momenta = atoms.get_momenta()
+    masses = atoms.get_masses()
+    velocity_com = momenta[indices].sum(axis=0) / masses[indices].sum()
 
     assert max(velocity_com) == pytest.approx(0.0, abs=1e-8)
 
@@ -49,7 +56,9 @@ def test_center_of_mass_velocity(atoms: Atoms):
 @pytest.mark.optimize
 def test_center_of_mass_force(atoms: Atoms):
     """Test if the corrected forces are along the COM-preserving direction."""
+    indices = [3, 4, 5]
     rnd = np.random.default_rng(42)
     forces = rnd.random(size=atoms.positions.shape)
-    FixCom().adjust_forces(atoms, forces)
-    np.testing.assert_allclose(atoms.get_masses() @ forces, 0.0, atol=1e-12)
+    FixSubsetCom(indices=indices).adjust_forces(atoms, forces)
+    tmp = atoms.get_masses()[indices] @ forces[indices]
+    np.testing.assert_allclose(tmp, 0.0, atol=1e-12)